Fixed compilation problems on Linux

+++ client/utils.h
@@ -13,13 +13,14 @@ inline QString getValue(const QString& data, const QString& key)
     return res;
 }
 
+
 inline QString GenerateUUID()
 {
     QTime now = QTime::currentTime();
     qsrand(now.msec());
 
     QString uuid(24 * 2);
-    for (size_t i = 0; i < 24 * 2; i++) {
+    for (int i = 0; i < 24 * 2; i++) {
         int c = qrand() % 16;
         Q_ASSERT(c >= 0);
         uuid[i] = c < 10 ? '0' + c : 'A' + (c - 10);
